Added cc:notice to the deed template
Antoine Isaac noticed that the rdf of the licenses was out of sync with
the rdfa of the deed pages.

To solve this Antoine and I propose to change the basic template of the
deed pages where we changed the moved the cc:Attribution requirement in
rdfa and added a cc:Notice requirement.

It was necessary to move the cc:Requires cc:Attribution due to the
placement of the cc:Requires cc:Notice tag. If we leave the cc:Requires
cc:Attribution as it was it implied that there was an hierarchy in
these statements, which there isn’t.