Update zero_deed.html
Should fix https://github.com/creativecommons/creativecommons.org/issues/864 as per suggestion of @robmyers 

Rationale. CC0 is not a license that is included in the license chooser. the license chooser API (that is being revamped as we speak I believe) does not know how to handle a license that does not exist in the chooser.

Fix is not ready for multilingual CC0 Deed pages. But better than having an error appear.